Actalizacion

+++ src/main/java/org/uv/Abarrotes/servicio/DetalleReporteService.java
@@ -6,10 +6,7 @@ package org.uv.Abarrotes.servicio;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.uv.Abarrotes.DTOs.DTODetalleReporte;
 import org.uv.Abarrotes.modelos.DetalleReporte;
-import org.uv.Abarrotes.modelos.DetalleVenta;
-import org.uv.Abarrotes.modelos.Reporte;
 import org.uv.Abarrotes.repositorio.DetalleReporteRepository;
 
 /**
@@ -22,12 +19,18 @@ public class DetalleReporteService {
     private DetalleReporteRepository detalleReporteRepository;
 
     public void borraDetalleReporte(Long id) {
-
-        detalleReporteRepository.deleteById(id);;
+        DetalleReporte detalleReporte = detalleReporteRepository.findById(id).orElse(null);
+        if (detalleReporte == null) {
+            detalleReporteRepository.deleteById(id);
+        }
+        
     }
 
     public DetalleReporte obtenerDetalleReportePorId(Long id) {
         return detalleReporteRepository.findById(id).orElse(null);
     }
 
+    public void borrarDetalleReportePorReporteId(Long reporteId) {
+        detalleReporteRepository.deleteByReporte_idReporte(reporteId);
+    }
 }
